Refactor theme system and update asset structure
Moved legacy and documentation files into MeVoy/MeVoy for better organization. Replaced base.css with new theme-based CSS files (mvclassic, cozy_blue, rose_warm) and updated imports in App.js. Updated ActionButton to use CSS variables for colors. Deleted unused assets and updated image paths. Added a test HTML for theme preview. Various code comments and minor cleanups for maintainability.
