+++ MeVoy/src/App.css
@@ -1,3 +1,8 @@
+
+/*
+  OUTDATED
+*/
+
 :root {
   --primary: #3b82f6;
   --primary-hover: #2563eb;
