+++ MeVoy/src/components/ActionBar.jsx
@@ -1,6 +1,8 @@
 import React from "react";
 
 /**
+ * ToDo: This...? This needs a LOT of work.
+ * 
  * Barra de acciones Editar/Guardar/Cancelar, separada y fija en desktop.
  * Props:
  * - editMode, onEdit, onSave, onCancel, guardado
