--- MeVoy/src/styles/base.css
@@ -1,92 +1,0 @@
-/* src/styles/base.css – tokens y estilos globales */
-
-:root {
-  /* Colores principales */
-  --color-bg:            #f8fafc;          /* gris‑azulado muy claro */
-  --color-card:          #ffffff;
-  --color-border:        #e2e8f0;          /* slate‑200 */
-  --color-text:          #222222;
-  --color-muted:         #555;
-  --color-primary:       #3b82f6;          /* blue‑500 */
-  --color-primary-hover: #2563eb;          /* blue‑600 */
-
-  /* Otros tokens */
-  --radius-lg:   1rem;
-  --radius-md:   .5rem;
-  --shadow-lg:   0 15px 35px rgba(0,0,0,.25);
-  --shadow-sm:   0 2px 6px rgba(0,0,0,.12);
-
-  font-family: "Inter", system-ui, sans-serif;
-}
-
-/* --- Reseteo básico ----------------------------------------------- */
-html,body { height: 100%; margin: 0; background: var(--color-bg); }
-
-h1,h2,h3,h4,h5 { color: var(--color-text); margin: 0 0 .5rem; }
-
-p { color: var(--color-muted); margin: .25rem 0; font-size: .95rem; }
-
-/* --- Botones ------------------------------------------------------- */
-button {
-  font: inherit;
-  cursor: pointer;
-  border: none;
-  transition: background .2s ease, opacity .2s ease, transform .15s ease;
-}
-
-.btn {
-  padding: .65rem 1rem;
-  border-radius: var(--radius-md);
-  font-weight: 600;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: .35rem;
-}
-
-.btn-primary {
-  background: var(--color-primary);
-  color: #fff;
-}
-.btn-primary:hover   { background: var(--color-primary-hover); }
-.btn-primary:disabled { opacity: .45; cursor: not-allowed; }
-
-.btn-secondary {
-  background: #e0e0e0;
-  color: #333;
-}
-.btn-secondary:hover { background: #d5d5d5; }
-
-/* --- Tarjetas ------------------------------------------------------ */
-.card {
-  background: var(--color-card);
-  border-radius: var(--radius-lg);
-  box-shadow: var(--shadow-sm);
-  padding: 1rem 1.25rem;
-}
-
-/* --- Inputs -------------------------------------------------------- */
-.input {
-  width: 100%;
-  padding: .65rem .85rem;
-  border: 1px solid var(--color-border);
-  border-radius: var(--radius-md);
-  font: inherit;
-  background: #fff;
-}
-.input:focus { outline: 2px solid var(--color-primary); outline-offset: 1px; }
-
-/* --- Utilidades ---------------------------------------------------- */
-.text-center { text-align: center; }
-.mt-1 { margin-top: .25rem; }
-.mt-2 { margin-top: .5rem; }
-.mt-4 { margin-top: 1rem; }
-.mb-2 { margin-bottom: .5rem; }
-.flex   { display: flex; }
-.flex-1 { flex: 1; }
-.gap-1  { gap: .25rem; }
-.gap-2  { gap: .5rem; }
-
-@media (max-width: 480px) {
-  .card { padding: .9rem 1rem; }
-}
