+++ MeVoy/src/styles/cozyglow/color_themes/cozy_blue.css
@@ -1,0 +1,92 @@
+:root {
+  /* === Primitives === */
+  /* Soft Sky Blue (Primary) */
+  --color-sky-50: #f2f6f8;
+  --color-sky-100: #d6e3ea;
+  --color-sky-200: #b3cbd5;
+  --color-sky-300: #8fb3c0;
+  --color-sky-400: #6c9bab;
+  --color-sky-500: #4e7f8f; /* Primary */
+  --color-sky-600: #3e6673;
+  --color-sky-700: #2f4e58;
+  --color-sky-800: #20363e;
+  --color-sky-900: #152429;
+
+  /* Warm Sand / Amber (Secondary) */
+  --color-sand-50: #faf6f2;
+  --color-sand-100: #f0e3d6;
+  --color-sand-200: #e2c9ad;
+  --color-sand-300: #d4b07a;
+  --color-sand-400: #c59646;
+  --color-sand-500: #a97b32; /* Secondary */
+  --color-sand-600: #896328;
+  --color-sand-700: #6a4c1e;
+  --color-sand-800: #4b3614;
+  --color-sand-900: #2d200a;
+
+  /* Neutrals (Dark Surfaces) */
+  --color-gray-50: #ececec;
+  --color-gray-100: #d2d2d4;
+  --color-gray-200: #b8b8bb;
+  --color-gray-300: #9e9ea3;
+  --color-gray-400: #84848a;
+  --color-gray-500: #6a6a70;
+  --color-gray-600: #515156;
+  --color-gray-700: #39393d;
+  --color-gray-800: #242427;
+  --color-gray-900: #18181a;
+
+  /* Status Colors (muted) */
+  --color-green-500: #4b7d64;
+  --color-green-600: #3c6450;
+  --color-yellow-500: #d2a752;
+  --color-yellow-600: #b28b40;
+  --color-red-500: #b35c5c;
+  --color-red-600: #914a4a;
+  --color-blue-500: #5b7fa3;
+  --color-blue-600: #476480;
+
+  /* === Semantic tokens === */
+  --color-primary: var(--color-sky-500);
+  --color-primary-hover: var(--color-sky-600);
+  --color-primary-active: var(--color-sky-700);
+
+  --color-secondary: var(--color-sand-500);
+  --color-secondary-hover: var(--color-sand-600);
+  --color-secondary-active: var(--color-sand-700);
+
+  --color-accent: var(--color-sand-400);
+  --color-accent-hover: var(--color-sand-500);
+
+  --color-success: var(--color-green-500);
+  --color-success-hover: var(--color-green-600);
+
+  --color-warning: var(--color-yellow-500);
+  --color-warning-hover: var(--color-yellow-600);
+
+  --color-danger: var(--color-red-500);
+  --color-danger-hover: var(--color-red-600);
+
+  --color-info: var(--color-blue-500);
+  --color-info-hover: var(--color-blue-600);
+
+  --color-surface: var(--color-gray-900);
+  --color-surface-alt: var(--color-gray-800);
+  --color-surface-hover: var(--color-gray-700);
+  --color-overlay: rgba(0,0,0,0.6);
+
+  --color-text-primary: #e6e9ea;
+  --color-text-secondary: #b3b6b8;
+  --color-text-disabled: #8d8f91;
+
+  --color-text-on-primary: #f6f9fa;
+  --color-text-on-secondary: #fffaf5;
+  --color-text-on-danger: #fff8f8;
+
+  --color-border: var(--color-gray-700);
+  --color-border-strong: var(--color-gray-500);
+  --color-divider: var(--color-gray-700);
+
+  --color-focus-ring: var(--color-sand-400);
+  --color-selection: var(--color-sky-300);
+}
