+++ MeVoy/src/styles/cozyglow/color_themes/mvclassic.css
@@ -1,0 +1,98 @@
+:root {
+  /* === Primitives === */
+  /* Blues (Primary) */
+  --color-blue-50: #f2f6fa;
+  --color-blue-100: #d9e4f4;
+  --color-blue-200: #b3caeb;
+  --color-blue-300: #8daee0;
+  --color-blue-400: #6a93d4;
+  --color-blue-500: #4a7bd1; 
+  --color-blue-600: #3c66ad;
+  --color-blue-700: #2f5289; /* Primary */
+  --color-blue-800: #223d64;
+  --color-blue-900: #162a44;
+
+  /* Greens (Secondary) */
+  --color-green-50: #f0faf7;
+  --color-green-100: #d6f1e8;
+  --color-green-200: #aee0d1;
+  --color-green-300: #86cfba;
+  --color-green-400: #63c0a7;
+  --color-green-500: #46b89d; 
+  --color-green-600: #3aa087;
+  --color-green-700: #2e7f6c; /* Secondary */
+  --color-green-800: #225f51;
+  --color-green-900: #153f36;
+
+  /* Neutrals */
+  --color-gray-50: #f9fafb;
+  --color-gray-100: #f3f4f6;
+  --color-gray-200: #e5e7eb;
+  --color-gray-300: #d1d5db;
+  --color-gray-400: #9ca3af;
+  --color-gray-500: #6b7280;
+  --color-gray-600: #4b5563;
+  --color-gray-700: #374151;
+  --color-gray-800: #1f2937;
+  --color-gray-900: #111827;
+
+  /* Status */
+  --color-success-500: #4c9d7a;
+  --color-warning-500: #d2a752;
+  --color-danger-500: #c46363;
+  --color-info-500: #5b8fbf;
+
+  /* === Light Theme Semantic === */
+  --color-primary: var(--color-blue-600);
+  --color-primary-hover: var(--color-blue-800);
+  --color-secondary: var(--color-green-600);
+  --color-secondary-hover: var(--color-green-800);
+
+  --color-success: var(--color-success-500);
+  --color-warning: var(--color-warning-500);
+  --color-danger: var(--color-danger-500);
+  --color-info: var(--color-info-500);
+
+  --color-surface: var(--color-gray-50);
+  --color-surface-alt: var(--color-gray-100);
+  --color-surface-hover: var(--color-gray-200);
+  --color-overlay: rgba(0,0,0,0.04);
+
+  --color-text-primary: var(--color-gray-900);
+  --color-text-secondary: var(--color-gray-600);
+  --color-text-disabled: var(--color-gray-400);
+  --color-text-on-primary: #ffffff;
+  --color-text-on-secondary: #ffffff;
+  --color-text-on-danger: #ffffff;
+
+  --color-border: var(--color-gray-300);
+  --color-border-strong: var(--color-gray-500);
+  --color-divider: var(--color-gray-200);
+
+  --color-focus-ring: var(--color-green-400);
+  --color-selection: var(--color-blue-100);
+}
+
+/* === Dark Theme === */
+@media (prefers-color-scheme: dark) {
+  :root {
+    --color-surface: #111827;
+    --color-surface-alt: #1b2537;
+    --color-surface-hover: #2b3650;
+    --color-overlay: rgba(0,0,0,0.6);
+
+    --color-text-primary: #e5e9f0;
+    --color-text-secondary: #b3b6b8;
+    --color-text-disabled: #8d8f91;
+    --color-text-on-primary: #f6f9fa;
+    --color-text-on-secondary: #f6fafa;
+    --color-text-on-danger: #fff8f8;
+
+    --color-border: #2f3b52;
+    --color-border-strong: #46536d;
+    --color-divider: #2f3b52;
+
+    --color-focus-ring: #58d5b2;
+    --color-selection: #3a5b8f;
+  }
+}
