+++ MeVoy/src/styles/cozyglow/color_themes/rose_warm.css
@@ -1,0 +1,92 @@
+:root {
+  /* === Primitives === */
+  /* Dusty Rose (Primary) */
+  --color-rose-50: #fdf5f7;
+  --color-rose-100: #f8dbe1;
+  --color-rose-200: #f1b8c7;
+  --color-rose-300: #e890a8;
+  --color-rose-400: #db6d8d;
+  --color-rose-500: #c94f73; /* Primary */
+  --color-rose-600: #a63f5f;
+  --color-rose-700: #84324d;
+  --color-rose-800: #62263a;
+  --color-rose-900: #411927;
+
+  /* Sage / Muted Teal (Secondary) */
+  --color-sage-50: #f2f9f8;
+  --color-sage-100: #d8ebe9;
+  --color-sage-200: #b8d9d5;
+  --color-sage-300: #93c5c0;
+  --color-sage-400: #6daaa4;
+  --color-sage-500: #4f8f89; /* Secondary */
+  --color-sage-600: #3f726d;
+  --color-sage-700: #315854;
+  --color-sage-800: #223f3b;
+  --color-sage-900: #152725;
+
+  /* Neutrals (Dark Surfaces) */
+  --color-gray-50: #eaeaea;
+  --color-gray-100: #cfcfd2;
+  --color-gray-200: #b3b3b8;
+  --color-gray-300: #98989f;
+  --color-gray-400: #7d7d86;
+  --color-gray-500: #62626b;
+  --color-gray-600: #4a4a52;
+  --color-gray-700: #33333a;
+  --color-gray-800: #1f1f25;
+  --color-gray-900: #141418;
+
+  /* Status Colors (Darkened & Softened) */
+  --color-green-500: #4c8c6e;
+  --color-green-600: #3b6e57;
+  --color-yellow-500: #d2a752;
+  --color-yellow-600: #b28b40;
+  --color-red-500: #c45555;
+  --color-red-600: #a34444;
+  --color-blue-500: #5b7fa3;
+  --color-blue-600: #476480;
+
+  /* === Semantic tokens === */
+  --color-primary: var(--color-rose-500);
+  --color-primary-hover: var(--color-rose-600);
+  --color-primary-active: var(--color-rose-700);
+
+  --color-secondary: var(--color-sage-500);
+  --color-secondary-hover: var(--color-sage-600);
+  --color-secondary-active: var(--color-sage-700);
+
+  --color-accent: var(--color-blue-500);
+  --color-accent-hover: var(--color-blue-600);
+
+  --color-success: var(--color-green-500);
+  --color-success-hover: var(--color-green-600);
+
+  --color-warning: var(--color-yellow-500);
+  --color-warning-hover: var(--color-yellow-600);
+
+  --color-danger: var(--color-red-500);
+  --color-danger-hover: var(--color-red-600);
+
+  --color-info: var(--color-blue-500);
+  --color-info-hover: var(--color-blue-600);
+
+  --color-surface: var(--color-gray-900);
+  --color-surface-alt: var(--color-gray-800);
+  --color-surface-hover: var(--color-gray-700);
+  --color-overlay: rgba(0,0,0,0.6);
+
+  --color-text-primary: #e5e5e5;
+  --color-text-secondary: #b3b3b8;
+  --color-text-disabled: #7d7d86;
+
+  --color-text-on-primary: #fffafc;
+  --color-text-on-secondary: #f6fafa;
+  --color-text-on-danger: #fff7f7;
+
+  --color-border: var(--color-gray-700);
+  --color-border-strong: var(--color-gray-500);
+  --color-divider: var(--color-gray-700);
+
+  --color-focus-ring: var(--color-sage-400);
+  --color-selection: var(--color-rose-300);
+}
